Create BurnWallet
This code defines a module called "Burnwallet" which provides a smart contract/wallet that can receive custom tokens, but not send them. 

The module defines a struct for storing custom tokens and a storage vector to hold them. It also defines two public functions: "deposit" for depositing tokens into the wallet and "add_token" for adding custom tokens to the storage. 

The "deposit" function checks if the token being deposited is recognized and if it is, it is deposited into the wallet. If the token is not recognized, it is added to the storage and then deposited. 

The "add_token" function checks if the token already exists in the storage and if it doesn't, it adds it. The module aims to provide a secure way to receive custom tokens without the risk of them being withdrawn from the wallet.